chore(fbal): Refactor Unit Tests (#276)
* feat(fbal): contract factory tests

* fix(ci): bad code formatting

* feat(fbal): refactor fbal builder tests

* chore: fix code formatting

---------
